old style rgb values are allowed to be floats
git-svn-id: https://fractalfrost.svn.sourceforge.net/svnroot/fractalfrost/trunk@598 24c11add-7f9c-4b28-9183-29c26a6ab64e

+++ fr0stlib/__init__.py
@@ -348,7 +348,7 @@ class Palette(collections.Sequence):
                 if color.get('rgba') is not None and color.get('rgb') is None:
                     raise ParsingError('Only rgb palettes are currently supported')
 
-                palette.data[int(color.get('index'))] = map(int, color.get('rgb').split())
+                palette.data[int(color.get('index'))] = map(float, color.get('rgb').split())
 
         return palette
